small refactoring
-created models directory for posts object, renamed method to view roll-dice view
